+++ src/bit_match_gen/__init__.py
@@ -1,1 +1,0 @@
-from bit_match_gen.main import main